+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 5.0.0-beta.11
 * Cross-compile Rutta for .net core, subject to these limitations https://github.com/Topshelf/Topshelf/issues/444
+* Fix #330 – tagging Gauges
 
 #### 5.0.0-beta.10
 * Bump dependency on NodaTime, Http.fs